internal documentation

Documentation for PhyloSummaries's internal functions. These functions are not exported and their access (API) should not be considered stable. But they can still be used, like this for example: PhyloSummaries.foo() for a function named foo().

functions & types

PhyloSummaries.consensus_bipartitions! — Method

consensus_bipartitions!(splitcounts::Dictionary{<:SplitTuple,Int},
    proportion::Number, numtrees::Number)

Filter dictionary splitcounts to keep only the entries whose frequency (count value in the dictionary) are greater than proportion × numtrees, or equal to numtrees when proportion is 1. Bipartitions with frequency weight over 50% must be compatible with each other. Bipartitions are retained one by one, from most to least frequent, so long as they are compatible with bipartititions previously kept.

The result can be passed to tree_from_bipartitions to construct the associated consensus tree topology. proportion = 1 corresponds to the strict consensus tree, proportion = 0.5 corresponds to the majority-rule consensus tree and proportion = 0 to a greedy consensus tree.

Output: splitcounts modified, with some entries filtered out, and sorted by frequency (from least to most frequent) if proportion<0.5.

Assumption: all counts are positive.

Example

julia> using Dictionaries

julia> bp = [(true,false), (false,false), (true,true)]; freq=(3,1,4);

julia> splitcounts = dictionary(zip(bp, freq))
3-element Dictionary{Tuple{Bool, Bool}, Int64}:
  (true, false) │ 3
 (false, false) │ 1
   (true, true) │ 4

julia> PhyloSummaries.consensus_bipartitions!(splitcounts, 0.5, 4)
2-element Dictionary{Tuple{Bool, Bool}, Int64}:
 (true, false) │ 3
  (true, true) │ 4

julia> splitcounts = dictionary(zip(bp, freq)); # reset as earlier

julia> PhyloSummaries.consensus_bipartitions!(splitcounts, 0, 4)
3-element Dictionary{Tuple{Bool, Bool}, Int64}:
 (false, false) │ 1
  (true, false) │ 3
   (true, true) │ 4

PhyloSummaries.count_bipartitions! — Method

count_bipartitions!(counts, net, taxa, rooted)

Count bipartitions in net, and add them to counts. If a new bipartition is found, a new key is added to counts with value 1. If net has a bipartition already present as a key in counts, then the corresponding value is incremented by 1. net is not modified.

By default, input trees are considered unrooted, and bipartitions are considered. Use rooted=true to consider all input trees as rooted, in which case clades (rather than bipartitions) are used to build the output rooted consensus tree.

Only non-trivial (not all 0s or all 1s) splits contribute to the count totals.

If the tip labels in net do not match those in taxa (as a set), then an error will be thrown indirectly (via PhyloNetworks.hardwiredclusters).

PhyloSummaries.tree_from_bipartitions — Method

tree_from_bipartitions(taxa::Vector{String},
    clusters::Dictionary{SplitTuple,<:Number},
    ntrees::Number,
    supportaslength::Bool)

Construct a consensus tree from a compatible set of cluster, as a PhyloNetworks.HybridNetwork object. Each cluster is represented as a tuple key b, and is given a weight: its value clusters[b] divided by ntrees. For each cluster, a node is added to the consensus tree, whose descendant taxa is the set taxa[i] for indices i such that b[i] is true. The cluster's weight is stored in the node's .fvalue –-which is fragile, if clusters are to be considered as unrooted bipartitions. The weight is also stored in fields of the node's parent edge: .length and .y.

Assumption: the input clusters are pairwise tree-compatible, which is the condition for them to be the clusters of a valid rooted tree.

Used by: consensustree

PhyloSummaries.treecompatible — Method

treecompatible(a::SplitTuple, b::SplitTuple)

true / false if two clusters a and b are / are not tree-compatible.

If A is the cluster of descendants of a (with true entries in a) and if B is the cluster of descendant of b, these 2 clusters are tree-compatible if there exists some tree that has both clusters. This can be checked by the condition: A∩B is empty, or A⊆B, or B⊆A.

PhyloSummaries.tuple_from_clustervector — Method

tuple_from_clustervector(cluster01vector, rooted)

Tuple of booleans t with t[i] true / false if taxa[i] does / does not belong in the hardwired cluster vector cluster01vector of 0/1 integers; or nothing if the cluster is trivial (all 0s or all 1s).

If rooted is false, then clusters are considered as bipartitions and the last taxon is used as an outgroup with a false entry. For example, clusters 0011 and 1100 represent the same bipartition, and both would return tuple (true,true,false,false).

index

PhyloSummaries.consensus_bipartitions!

PhyloSummaries.count_bipartitions!

PhyloSummaries.tree_from_bipartitions

PhyloSummaries.treecompatible

PhyloSummaries.tuple_from_clustervector