public documentation

Documentation for PhyloSummaries's public (exported) functions. Most functions are internal (not exported).

functions & types

PhyloSummaries.consensustree — Method

consensustree(trees::AbstractVector{PN.HybridNetwork};
              rooted=false,
              proportion=0,
              supportaslength=false)

Consensus tree summarizing the bipartitions (or clades) shared by more than the required proportion of input trees. An ArgumentError is thrown if one input network is not a tree, or the list of input trees is empty, or if the input trees do not all have the same tip labels. Input trees are not modified.

Output: consensus tree as an object of type HybridNetwork.

The bipartition (or clade) support values are stored

as edge length with option supportaslength=true (not by default).

in the field .y of each internal edge (as external edges correspond to trivial bipartitions, which must necessarily be in all sampled trees). This field is used by writenewick to write edge support, with its option support=true. However, this .y field is internal, so it can be modified by other functions and should not be relied upon.

Support values are also stored in each node .fvalue for the clade descendant from that node. Note that storing support values at nodes is unsafe for unrooted bipartitions, as the bipartition associated with the node's descendants given some rooting may become one of node's child subclade after re-rooting.

By default, input trees are considered unrooted, and bipartitions are considered. Use rooted=true to consider all input trees as rooted, in which case clades (rather than bipartitions) are used to build the output rooted consensus tree.

By default, the greedy consensus consensus is calculated: the tree is built from the bipartitions (or clades) with the highest support, until no more can be added. The majority-rule tree can be obtained by using proportion=0.5: it is built only from bipartitions (or clades) present in more than 50% of the input trees.

assumptions and warnings:

Input trees are assumed to have their edges correctly directed. If unsure, run directedges!.(trees) prior.

Input trees should not have degree-2 nodes other than the root (nodes with 1 only parent and 1 child). If unsure, run removedegree2nodes!.(trees, true)) to keep their root even of degree 2 or removedegree2nodes!.(trees, false)) unroot them also.

fixit: make a future version summarize edge lengths in input trees and store their average in the consensus tree.

example

julia> nwk = ["((c,d),((a1,a2),b));", "(((a2,a1),b),c,d);", "(((a1,a2),c),d,b);"];

julia> treesample = readnewick.(nwk);

julia> con = consensustree(treesample); writenewick(con, round=true, support=true)
"(c,d,(b,(a1,a2)::1.0)::0.667);"

julia> con = consensustree(treesample; rooted=true); # greedy consensus

julia> writenewick(con, round=true, support=true)
"((d,c)::0.333,(b,(a1,a2)::1.0)::0.667);"

julia> [e.number => round(e.y, digits=3) for e in con.edge if !isexternal(e)] # edge number -> support
3-element Vector{Pair{Int64, Float64}}:
 6 => 0.333
 7 => 0.667
 8 => 1.0

julia> con = consensustree(treesample; rooted=true, proportion=0.5); # majority-rule

julia> writenewick(con, round=true, support=true)
"(c,d,(b,(a1,a2)::1.0)::0.667);"

julia> con = consensustree(treesample; proportion=0.75, supportaslength=true) |> writenewick
"(b,c,d,(a2,a1):1.0);"

index

PhyloSummaries.consensustree